rename & fix: enum클래스를 state패키지로 이동 & 판매자 구독 조회 시 currentPage 반환값 수정
