+++ src/main/java/com/leeforgiveness/memberservice/subscribe/state/SubscribeState.java
@@ -1,4 +1,4 @@
-package com.leeforgiveness.memberservice.subscribe;
+package com.leeforgiveness.memberservice.subscribe.state;
 
 public enum SubscribeState {
     SUBSCRIBE,
